Spark Effect
Added a spark effect to the rockets which was accomplished by adding two
new classes - Sparks and Particle.
Sparks is a class which allows customization of the range of sparks, the
force of the sparks and the intensity, it was implemented to control the
behaviour of individual particles.
Particle is a helper class for Sparks which recieves a location, colour,
speed and direction.
Sparks are managed by a queue with a limit of 15 Sparks per Rocket at any
given time.