Fixed Rotation Issue and Refractored
Added a series of configuration options to allow the rocket visualization
to be modified and displayed as specified.
Added different variations of the Target class to allow for different
types of targets.
Reformatted, refractored and commented existing code.
Fixed the issue with rotation of the rocket not matching the direction the
rocket was heading by reimplementing the code for controlling the rockets
movement.
Fixed the issue where rockets were not removed once they reached their
destination.
Added a MouseTarget class which makes the rockets target the mouse cursor.
Added a RandomTarget class which randomly targets the rocket elsewhere.
Updated the README.md to reflect the new experiment.