Loader Upgrade
Upgraded the loader index.html file to allow for rapid configuration
modifications and experiment switching.
Global configuration options are now loaded from the index page using a
combination of the default config in files.js and modified configuaration
specified by a form on the index page.
The index page allows selection of the experiment and altering of the
values for config.
The files.js file contains the default configuration for each experiment.
The README.md has been updated to include the 3D Picture experiment.
The configuration for each project is loaded in the preload function
through a function called loadConfig, any required additions to the
configuration are then performed.