Dotter
Started a new experiment called Dotter which consists of a series of dots
scattered randomly being joined by a line which then pulls the dots
together.
Dots randomly fluctuate in size, when a dot reaches a certain size it will
decrease to the minimum size and produce child dots.
Dots are coloured using Perlin Noise which gives some interesting visual
effects.
When a mouse is moved into the screen the dots will move away from the
mouse.